# aurora-go/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_aurora-go_aurora_internal_vo.TalkVO:
    properties:
      content:
        description: 说说内容不能为空
        type: string
      id:
        type: integer
      images:
        description: 说说图片JSON数组字符串
        type: string
      isTop:
        description: 置顶状态(0否 1是)，去掉required避免0值校验失败
        type: integer
      status:
        description: 说说状态(1公开 2私密)，去掉required避免0值校验失败
        type: integer
    required:
    - content
    type: object
  github_com_aurora-go_aurora_internal_vo.WebsiteConfigVO:
    properties:
      alipayQRCode:
        description: 对标Java：支付宝二维码
        type: string
      author:
        description: 对标Java：网站作者
        type: string
      authorAvatar:
        description: 对标Java：网站头像
        type: string
      authorIntro:
        description: 对标Java：网站作者介绍
        type: string
      beianNumber:
        description: 对标Java：工信部备案号
        type: string
      csdn:
        description: 对标Java：csdn
        type: string
      englishName:
        description: 对标Java：网站英文名称
        type: string
      favicon:
        description: 对标Java：favicon
        type: string
      gitee:
        description: 对标Java：gitee
        type: string
      github:
        description: 对标Java：github
        type: string
      gonganBeianNumber:
        description: 对标Java：公安部备案编号
        type: string
      isCommentReview:
        description: 对标Java：Integer（0关闭1开启）
        type: integer
      isEmailNotice:
        description: 对标Java：Integer（0关闭1开启）
        type: integer
      isReward:
        description: 对标Java：Integer（0关闭1开启）
        type: integer
      juejin:
        description: 对标Java：juejin
        type: string
      logo:
        description: 对标Java：网站logo
        type: string
      multiLanguage:
        description: 对标Java：Integer（0关闭1开启）
        type: integer
      name:
        description: 对标Java：网站名称
        type: string
      notice:
        description: 对标Java：网站公告
        type: string
      qq:
        description: 对标Java：qq
        type: string
      qqLogin:
        description: 对标Java：Integer（0关闭1开启）
        type: integer
      stackoverflow:
        description: 对标Java：stackoverflow
        type: string
      touristAvatar:
        description: 对标Java：游客头像
        type: string
      twitter:
        description: 对标Java：twitter
        type: string
      userAvatar:
        description: 对标Java：用户头像
        type: string
      weChat:
        description: 对标Java：微信
        type: string
      websiteCreateTime:
        description: 对标Java：网站创建时间
        type: string
      websiteTitle:
        description: 对标Java：网页标题
        type: string
      weiXinQRCode:
        description: 对标Java：微信二维码
        type: string
      weibo:
        description: 对标Java：微博
        type: string
      zhihu:
        description: 对标Java：zhihu
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: aurora@example.com
    name: Aurora Team
  description: Aurora 博客系统 Go 1.26 后端 - 基于 tRPC-Agent-Go 的 AI 驱动博客平台
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/nicepkg/aurora
  title: Aurora Blog API
  version: "1.0"
paths:
...
  /api/admin/talks:
    delete:
      consumes:
      - application/json
      description: 后台批量删除说说
      parameters:
      - description: 说说ID列表
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 批量删除说说
      tags:
      - 说说
    get:
      consumes:
      - application/json
      description: 后台获取说说列表（分页）
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 当前页码
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回说说列表
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 后台说说列表
      tags:
      - 说说
    post:
      consumes:
      - application/json
      description: 后台保存或更新说说
      parameters:
      - description: 说说内容
        in: body
        name: talkVO
        required: true
        schema:
          $ref: '#/definitions/github_com_aurora-go_aurora_internal_vo.TalkVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 保存或更新说说
      tags:
      - 说说
  /api/admin/talks/{id}:
    get:
      consumes:
      - application/json
      description: 后台根据ID获取说说详情
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回说说详情
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 后台获取说说详情
      tags:
      - 说说
  /api/admin/talks/images:
    post:
      consumes:
      - multipart/form-data
      description: 上传说说中的图片
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回图片URL
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 上传说说图片
      tags:
      - 说说
...
  /api/admin/website/config:
    get:
      consumes:
      - application/json
      description: 获取网站前台公开配置
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回网站配置
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: 获取网站前台配置
      tags:
      - 网站配置
    put:
      consumes:
      - application/json
      description: 后台更新网站配置
      parameters:
      - description: 网站配置信息
        in: body
        name: configVO
        required: true
        schema:
          $ref: '#/definitions/github_com_aurora-go_aurora_internal_vo.WebsiteConfigVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 更新网站配置
      tags:
      - 网站配置
  /api/admin/website/config/images:
    post:
      consumes:
      - multipart/form-data
      description: 上传网站配置图片（Logo/Favicon/头像等）
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      - description: 图片类型
        in: formData
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回图片URL
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 上传网站图片
      tags:
      - 网站配置
...
  /api/talks:
    get:
      consumes:
      - application/json
      description: 获取前台说说列表（分页，按时间倒序）
      parameters:
      - description: 当前页码
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回说说列表
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: 获取说说列表
      tags:
      - 说说
  /api/talks/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取说说详情
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回说说详情
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: 获取说说详情
      tags:
      - 说说
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Token 认证，格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"